feat(cli): tier-3 commands — fmt, explain, history, run --output json (#387,#389,#390,#391)
Four tier-3 CLI ergonomics features (implementation picked up from a prior
session, then finished: fixed the test build, added coverage, docs).

- #387 `faucet fmt [--check|--stdout]`: canonicalize a config (stable key order),
  idempotent; `--check` is a CI gate. Comments not preserved.
- #389 `faucet explain [--json|--rows]`: plain-English narration of a pipeline
  (source→transforms→sink, matrix, delivery). Fully offline; never prints secrets.
- #390 `faucet run --output <text|json|ndjson>`: machine-readable end-of-run
  summary (per-row + totals) with clean stdout for CI/cron/Slack.
- #391 `faucet history [--limit|--row|--json]`: terminal view of the `catalog:`
  store's run history (status/duration/throughput). Requires the `catalog` feature.

Finishing work: added the new `output` field to the two RunArgs test literals
(lineage_run, plan_diff_e2e); extracted a pure `select_runs` in history + a
no-catalog error test; added `run()`-flow tests for explain; a MinIO-free sqlite
`history_e2e` integration test seeding run records. Local coverage: fmt 93%,
explain 92%, history 96% (with e2e), run added ~92%. clippy + fmt clean.

Closes #387. Closes #389. Closes #390. Closes #391.

+++ cli/src/executor.rs
@@ -1469,13 +1469,17 @@ async fn run_one_invocation(
     // is only known when the source sampler was installed (a `lineage:` or
     // `catalog:` block); otherwise it stays `None` rather than guessing.
     #[cfg(feature = "lineage")]
-    let records_read = in_sample.as_ref().map(|s| s.count() as u64);
+    let records_read = in_sample.as_ref().map(|s| s.count());
     #[cfg(not(feature = "lineage"))]
     let records_read: Option<u64> = None;
     let stats = PipelineStats {
         records_written: result.records_written,
         records_read,
-        dlq_count: result.dlq.as_ref().map(|d| d.records_dlq as u64).unwrap_or(0),
+        dlq_count: result
+            .dlq
+            .as_ref()
+            .map(|d| d.records_dlq as u64)
+            .unwrap_or(0),
         bookmark: result.bookmark.clone(),
     };
